Created homepage test
