--- tests/test_view_source.py
@@ -1,39 +1,0 @@
-#!/usr/bin/env python
-
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-import pytest
-from unittestzero import Assert
-
-from pages.home import HomePage
-from pages.log_in_or_create_account import LogInOrCreateAccountPage
-
-
-class TestViewSource:
-
-    @pytest.mark.nondestructive
-    def test_visitor_can_view_source(self, mozwebqa):
-        home_pg = HomePage(mozwebqa)
-        home_pg.go_to_home_page()
-
-        view_source_pg = home_pg.header_region.click_view_source()
-        Assert.true(view_source_pg.is_the_current_page)
-        Assert.greater(len(view_source_pg.source_textarea.strip()), 0)
-
-    @pytest.mark.nondestructive
-    def test_logged_in_user_can_view_source(self, mozwebqa):
-        home_pg = HomePage(mozwebqa)
-        home_pg.go_to_home_page()
-
-        home_pg.personal_tools_region.click_log_in_or_create_account()
-        log_in_or_create_account_pg = LogInOrCreateAccountPage(mozwebqa)
-        log_in_or_create_account_pg.log_in()
-
-        Assert.false(home_pg.personal_tools_region.is_log_in_or_create_account_visible)
-        Assert.true(home_pg.personal_tools_region.is_log_out_visible)
-
-        view_source_pg = home_pg.header_region.click_view_source()
-        Assert.true(view_source_pg.is_the_current_page)
-        Assert.greater(len(view_source_pg.source_textarea.strip()), 0)
